Moved ApplicationController code for multisite into a concern
