+++ app/models/guts/metafield.rb
@@ -1,7 +1,7 @@
 module Guts
   # Metafield model
   class Metafield < ActiveRecord::Base
-    include SiteScopeConcern
+    include MultisiteScopeConcern
     
     belongs_to :site
     belongs_to :fieldable, polymorphic: true, required: false
