+++ app/models/guts/option.rb
@@ -2,7 +2,7 @@ module Guts
   # Option model
   class Option < ActiveRecord::Base
     include TrackableConcern
-    include SiteScopeConcern
+    include MultisiteScopeConcern
     
     # Regex for replacing key values with
     KEY_REGEX = /[^a-zA-Z0-9_ ]/i
