+++ app/models/guts/tracker.rb
@@ -1,7 +1,7 @@
 module Guts
   # Tracker model
   class Tracker < ActiveRecord::Base
-    include SiteScopeConcern
+    include MultisiteScopeConcern
     
     belongs_to :site
     belongs_to :object, polymorphic: true
